requirements.txt
astroid==2.3.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
httplib2==0.14.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
oauthlib==3.1.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pylint==2.4.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.6
wrapt==1.11.2
